Random number testing
This project relies heavily on repeatable number generation. At the moment I believe it will be best for the player class to retain an array of integers and doubles for the various generations necessary. This may or may not be the final route but this ensures consistency at the very least.