Added biased amiability list
Used a binomial distribution centered around Neutral. Values get shifted based on the passed amiability level. Hostile would map to only Hostile, Unfriendly, and Neutral, while VeryFriendly would map to VF, F, and N. If set to Neutral there is a nice, distribution across all with the most being in Neutral. This could alternatively have been done with a Beta distribution where the output was put into buckets to keep a few individuals in the extremes but I think this keeps the "community" consistent.