Revamped racemakeup, minor corrections elsewhere
RaceMakup had an overhaul. I wanted random races to be the predominant race and an appropriate "normal-like" distribution to represent the races in a region. To do that the races underwent a shuffling. A random "normal-like" distribution was then assigned to the shuffled races. Minor corrections were made to flora/fauna. Not sure how culture will affect things yet, nor natural resources.